+++ packages/react-router/src/Routes.tsx
@@ -1,0 +1,45 @@
+import React from "react";
+import { RouteProps } from "~/Route";
+import { Location } from "history";
+import { Route, Routes as BaseRoutes, RoutesProps as BaseRoutesProps } from "react-router-dom";
+import { useLocation } from "~/index";
+
+const createNativeRoute = (props: RouteProps, index: number, location: Location): JSX.Element => {
+    const newProps = { ...props };
+    if (!props.exact) {
+        newProps.path = `${props.path}/*`;
+    }
+    delete newProps["exact"];
+
+    if (typeof props.render === "function") {
+        newProps.element = props.render({ location });
+    } else if (props.component) {
+        const Component = props.component;
+        newProps.element = <Component />;
+    }
+    delete newProps["render"];
+    delete newProps["component"];
+
+    if (props.children) {
+        newProps.element = props.children;
+    }
+
+    delete newProps["children"];
+
+    return <Route key={`${newProps.path}:${index}`} {...newProps} />;
+};
+
+export interface RoutesProps extends BaseRoutesProps {}
+
+export const Routes: React.FC<RoutesProps> = props => {
+    const location = useLocation();
+
+    const children = React.Children.map(props.children, (route: any, index) => {
+        if (!route) {
+            return null;
+        }
+        return createNativeRoute(route.props, index, location);
+    });
+
+    return <BaseRoutes>{children}</BaseRoutes>;
+};
